refactor profile
