+++ crates/notedeck_columns/src/ui/note/mod.rs
@@ -25,7 +25,7 @@ use enostr::{NoteId, Pubkey};
 use nostrdb::{Ndb, Note, NoteKey, Transaction};
 use notedeck::{CachedNote, ImageCache, NoteCache, NotedeckTextStyle};
 
-use super::profile::preview::{get_display_name, one_line_display_name_widget};
+use super::profile::{get_display_name, preview::one_line_display_name_widget};
 
 pub struct NoteView<'a> {
     ndb: &'a Ndb,
